backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
httpx==0.27.0
numpy==1.26.4
openpyxl==3.1.5
reportlab==4.2.0
pydantic==2.9.0
